trying to get this ish done

+++ backend/db/queries/delete_book_by_id.sql
@@ -1,2 +1,2 @@
 -- name: DeleteBookByID :exec
-DELETE FROM books WHERE id = $1;
+DELETE FROM books WHERE id = $1 AND user_id = $2;
